Creation of the pre-screening function Step0_LoopLineFilter.
The objective of this function is to analyze the loops that follow the direct line connecting any two punctures.  Generally it has been the case that invariant puncture sets consist of closely packed punctures which move as a group but do not intermix.  The hope is that analysis of these "loop-line" will accelerate the rest of the algorithm.