Avoid entropy overflow by rescaling length as we go (non-MEX version).
A bit kludgy: pass an extra flag saying we want the log-rescaling.
This flag has to be propagated from mtimes to loopsigma.  Right now
the rescaling is done using a simple L^1 norm, but probably it should
use lenfun.  Also, I changed the way intaxis is used, but hopefully
it's equivalent.  Doublecheck.  Also need to do the C++ implementation.